+++ ci4/app/Entities/DeploymentSpecificationDeploymentAnnotation.php
@@ -7,13 +7,15 @@ use App\Core\Entity;
  * @package App\Entities
  * @property string $name
  * @property string $value
+ * @property string $level
  * @property int $deployment_specification_id
  * @property DeploymentSpecification $deployment_specification
  */
 class DeploymentSpecificationDeploymentAnnotation extends Entity {
 
-    public static function Create(string $name, string $value): DeploymentSpecificationDeploymentAnnotation {
+    public static function Create(string $level, string $name, string $value): DeploymentSpecificationDeploymentAnnotation {
         $item = new DeploymentSpecificationDeploymentAnnotation();
+        $item->level = $level;
         $item->name = $name;
         $item->value = $value;
         $item->save();