+++ ci4/app/Interfaces/DeploymentAnnotation.php
@@ -1,5 +1,6 @@
 <?php namespace App\Interfaces;
 /**
+ * @property string $level
  * @property string $name
  * @property string $value
  */