+++ CHANGELOG.md
@@ -1,12 +1,26 @@
 # Changelog
 
+## v1.5.7 (2025-08-05)
+
+### Fixed bugs
+*
+
+### Enhancements
+* Annotations can now be set at both deployment level and pod level
+
+### Upgrade guide
+1. Deploy new image
+2. Migrate database [(Guide)](https://github.com/4spacesdk/kubernetes-service-orchestrator?tab=readme-ov-file#migrate-database-helm)
+
+
+
 ## v1.5.6 (2025-06-20)
 
 ### Fixed bugs
 *
 
 ### Enhancements
-* InitContainers can now be added to a migration job 
+* InitContainers can now be added to a migration job
 
 ### Upgrade guide
 1. Deploy new image